Move AuthStore to useAuth
